# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Library API documentation",
        "contact": {
            "email": "sferaved@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://library/api",
            "description": "Laravel Swagger API server"
        }
    ],
    "paths": {
        "/books": {
            "get": {
                "tags": [
                    "Books ALL"
                ],
                "summary": "Display a listing of the resource",
                "operationId": "books",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine"
                    },
                    "404": {
                        "description": "Example not found"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Books Creat"
                ],
                "summary": "Create yet another book record",
                "operationId": "bookCreate",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoreBookRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        },
        "/books/{id}": {
            "get": {
                "tags": [
                    "Books Find by id"
                ],
                "summary": "Get book by ID",
                "operationId": "booksGet",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of book",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/FindBookRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Books Update"
                ],
                "summary": "Update book by ID",
                "operationId": "booksUpdate",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of example",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/StoreBookRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Everything is fine",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/StoreBookRequest"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Books Destroy"
                ],
                "summary": "Delete book by ID",
                "operationId": "booksDestroy",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "The ID of example",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": "1"
                    }
                ],
                "responses": {
                    "202": {
                        "description": "Deleted"
                    }
                },
                "security": [
                    {
                        "api_key": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "FindBookRequest": {
                "title": "Library finding request",
                "description": "Some simple request createa as example",
                "properties": {
                    "id": {
                        "title": "Id",
                        "description": "Id for finding",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
            "StoreBookRequest": {
                "title": "Library storing request",
                "description": "Some simple request createa as example",
                "properties": {
                    "name": {
                        "title": "Name",
                        "description": "Name of key for storring",
                        "type": "string",
                        "example": "random"
                    },
                    "author_id": {
                        "title": "Author_id",
                        "description": "Author_id for storring",
                        "type": "integer",
                        "example": "1"
                    },
                    "description": {
                        "title": "Description",
                        "description": "Value for storring",
                        "type": "string",
                        "example": "awesome"
                    },
                    "cover": {
                        "title": "Cover",
                        "description": "Cover for storring",
                        "type": "string",
                        "example": "images/books/My new book.jpg"
                    },
                    "category_id": {
                        "title": "Category_id",
                        "description": "Author_id for storring",
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "X-APP-ID": {
                "type": "apiKey",
                "name": "X-APP-ID",
                "in": "header"
            }
        }
    },
    "tags": [
        {
            "name": "Library",
            "description": "Books page"
        }
    ]
}